requirements.txt
Pillow==10.0.0
qrcode==7.4.2
